+++ core/parser/platform/xiaoheihe.py
@@ -17,9 +17,9 @@ import aiohttp
 
 from ...logger import logger
 
-from .base import BaseVideoParser
-from ..utils import build_request_headers
 from ...constants import Config
+from ..utils import build_request_headers
+from .base import BaseVideoParser
 
 try:
     from cryptography.hazmat.decrepit.ciphers.algorithms import TripleDES