+++ core/parser/platform/xueqiu.py
@@ -10,8 +10,9 @@ from urllib.parse import urlparse
 
 import aiohttp
 
-from ...constants import Config
 from ...logger import logger
+
+from ...constants import Config
 from ...types import MediaMetadata
 from ..utils import build_request_headers
 from .base import BaseVideoParser